Organize checkpoint workflow resolvers
